requirements.txt
# Bugfixer Service Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
GitPython==3.1.40
openai==1.3.7
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
httpx==0.25.2
psutil==5.9.6
watchdog==3.0.0
sqlalchemy==2.0.23
alembic==1.12.1
